Start adding docs to MastodonClient trait
Also moves it to it's own module, to clean up src/lib.rs a bit
